[REFACTOR] Moved stuff that are profile related to users/profile
