+++ Eventy/app/src/main/res/layout/fragment_my_cards.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".users.MyCardsFragment">
+    tools:context=".users.profile.MyCardsFragment">
 
     <FrameLayout
         android:id="@+id/fragmentContainer"
