+++ Eventy/app/src/main/res/layout/fragment_organizer_events.xml
@@ -7,7 +7,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:orientation="vertical"
     android:background="@color/background_color"
-    tools:context=".users.OrganizerEventsFragment">
+    tools:context=".users.profile.OrganizerEventsFragment">
 
     <LinearLayout
         android:id="@+id/search_and_filter_container"
