+++ Eventy/app/src/main/res/layout/fragment_other_user_profile_page.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background_color"
-    tools:context=".users.OtherUserProfilePageFragment">
+    tools:context=".users.profile.OtherUserProfilePageFragment">
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
