+++ Eventy/app/src/main/res/navigation/mobile_navigation.xml
@@ -73,7 +73,7 @@
   
     <fragment
         android:id="@+id/nav_other_user_profile_page"
-        android:name="com.example.eventy.users.OtherUserProfilePageFragment"
+        android:name="com.example.eventy.users.profile.OtherUserProfilePageFragment"
         android:label="Other User Profile Page"
         tools:layout="@layout/fragment_other_user_profile_page"/>
